requirements.txt
sentence-transformers
faiss-cpu
langchain
langchain-ollama
